debug logic was separated from visitor
